requirements.txt
pymavlink==2.4.49
w1thermsensor==2.3.0
python-dotenv==1.0.1
Pillow>=10.0.0
